learn about useFrame
this is the one we use rotate the object we use

useFrame provides two arguments:

State → An object containing useful properties like camera, clock, and scene.
Delta → The time difference between frames (helps with smooth animations).
In your code, we don't need the state object, so we use _ to ignore it.

This is just a coding convention to indicate "I don't need this value."

ref.current.position.y=Math.sin(state.clock.getElapsedTime()*2)=This line makes the object move up and down
*2 means speed the wave